+++ app/build.gradle
@@ -5,7 +5,7 @@ android {
     buildToolsVersion "23.0.3"
 
     defaultConfig {
-        applicationId "nz.allthethings.whareora"
+        applicationId "nz.allthethings.wharehauora"
         minSdkVersion 21
         targetSdkVersion 23
         versionCode 1
