+++ app/src/main/res/layout/app_bar_main.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context="nz.allthethings.whareora.MainActivity">
+    tools:context="nz.allthethings.wharehauora.MainActivity">
 
     <android.support.design.widget.AppBarLayout
         android:layout_width="match_parent"
