+++ app/src/main/res/layout/content_main.xml
@@ -9,7 +9,7 @@
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context="nz.allthethings.whareora.MainActivity"
+    tools:context="nz.allthethings.wharehauora.MainActivity"
     tools:showIn="@layout/app_bar_main">
