+++ app/src/test/java/nz/allthethings/wharehauora/ExampleUnitTest.java
@@ -1,4 +1,4 @@
-package nz.allthethings.whareora;
+package nz.allthethings.wharehauora;
 
 import org.junit.Test;
